requirements.txt
streamlit>=1.32.0
requests>=2.31.0
beautifulsoup4>=4.12.3
lxml>=5.1.0
google-generativeai>=0.5.0
jinja2>=3.1.3
playwright>=1.43.0
